Use only Origin: field to signal APT preferences
Give preference to APT sources with "Pragmatux" set in the Origin: field
of their Release file. This permits overriding a package delivered by,
e.g., Debian, without worrying that a subsequent update to the package
in Debian's repository will replace the overidden package due to a
favorable version comparison.

The Label: field was also used for this purpose in earlier versions of
Pragmatux, but that use was depreciated. Remove that now.