Add a distribution suffix to the APT sources package
Because the APT sources package is manually handled, as the
starting point for installing the Pragmatux SDK, it's helpful to
it clear which version of Pragmatux it installs.

Note this does not imply a given distribution of the SDK can be
installed on any distribution of Debian. A SDK can be installed
only on the distribution of Debian from which that version of
Pragmatux is derived. This is a consequence of the deep
integation with the package management system.